requirements.txt
flask
os
pytesseract
pdf2image
Pillow
asyncio
pypdfium2

